+++ internal/repository/brand/interface.go
@@ -8,8 +8,8 @@ import (
 
 type Repository interface {
 	Create(ctx context.Context, brand models.Brand) error
-	Delete(ctx context.Context, brandID int) error
-	GetByID(ctx context.Context, brandID int) (*models.Brand, error)
+	Delete(ctx context.Context, brandID uint16) error
+	GetByID(ctx context.Context, brandID uint16) (*models.Brand, error)
 	Update(ctx context.Context, brand models.Brand) error
 	GetAll(ctx context.Context) ([]models.Brand, error)
 	InsertOrUpdate(ctx context.Context, brand models.Brand) error
